refactor(handlers): restructure the handlers package to have more clean files
